feat(apps/client/composables): extract useEscapeBehavior from BaseModal
